requirements.txt
yt-dlp>=2023.11.16
openai-whisper
deep-translator
torch
